- Changed injection to happen on OnAttached function  - deleted login fragment  - and moved RecorderExplanationFragment to recorder package  - Updated dagger library
