+++ app/build.gradle
@@ -102,7 +102,7 @@ dependencies {
     //dagger
     implementation 'com.google.dagger:dagger:2.25.2'
     kapt 'com.google.dagger:dagger-compiler:2.25.2'
-    implementation 'com.google.dagger:dagger-android-support:2.25.2'
+    implementation 'com.google.dagger:dagger-android-support:2.27'
     // if you use the support libraries
 
     // permission
