+++ app/src/main/java/com/example/coronadiagnosticapp/ui/fragments/information/InformationFragment.kt
@@ -1,5 +1,6 @@
 package com.example.coronadiagnosticapp.ui.fragments.information
 
+import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -28,15 +29,15 @@ class InformationFragment : ScopedFragment() {
     @Inject
     lateinit var viewModel: InformationViewModel
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        context?.getAppComponent()?.inject(this)
-    }
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? = inflater.inflate(R.layout.information_fragment, container, false)
+
+    override fun onAttach(context: Context) {
+        context.getAppComponent().inject(this)
+        super.onAttach(context)
+    }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
