--- app/src/main/java/com/example/coronadiagnosticapp/ui/fragments/login/LoginFragment.kt
@@ -1,25 +1,0 @@
-package com.example.coronadiagnosticapp.ui.fragments.login
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.coronadiagnosticapp.R
-
-class LoginFragment : Fragment() {
-
-    companion object {
-        fun newInstance() = LoginFragment()
-    }
-
-    private lateinit var viewModel: LoginViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.login_fragment, container, false)
-    }
-
-}
