+++ app/src/main/java/com/example/coronadiagnosticapp/ui/fragments/notification/NotificationFragment.kt
@@ -1,5 +1,6 @@
 package com.example.coronadiagnosticapp.ui.fragments.notification
 
+import android.content.Context
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
@@ -25,9 +26,9 @@ class NotificationFragment : Fragment() {
         savedInstanceState: Bundle?
     ) = inflater.inflate(R.layout.fragment_notification, container, false)
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        context?.getAppComponent()?.inject(this)
+    override fun onAttach(context: Context) {
+        context.getAppComponent().inject(this)
+        super.onAttach(context)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
