+++ app/src/main/java/com/example/coronadiagnosticapp/ui/fragments/recorder/RecorderExplanationFragment.kt
@@ -1,4 +1,4 @@
-package com.example.coronadiagnosticapp
+package com.example.coronadiagnosticapp.ui.fragments.recorder
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import com.example.coronadiagnosticapp.R
 import kotlinx.android.synthetic.main.fragment_recorder_explanation.*
