--- app/src/main/java/com/example/coronadiagnosticapp/ui/login/LoginViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.coronadiagnosticapp.ui.login
-
-import androidx.lifecycle.ViewModel
-
-class LoginViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
