+++ app/src/main/res/layout/fragment_recorder_explanation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RecorderExplanationFragment">
+    tools:context=".ui.fragments.recorder.RecorderExplanationFragment">
 
     <ScrollView
         android:layout_width="match_parent"
