+++ app/src/main/res/navigation/nav_graph.xml
@@ -86,7 +86,7 @@
     </fragment>
     <fragment
         android:id="@+id/recorderExplanation"
-        android:name="com.example.coronadiagnosticapp.RecorderExplanationFragment"
+        android:name="com.example.coronadiagnosticapp.ui.fragments.recorder.RecorderExplanationFragment"
         android:label="fragment_recorder_explanation"
         tools:layout="@layout/fragment_recorder_explanation">
         <action
